refactor(installations): remove deprecated namespaced API
Modular-only surface via getOrCreateModularInstance with multi-app
support; drop namespaced registry and deprecation proxy. Document removal
in v26 migration guide.
Modular-only entry in lib/index.ts (no lib/modular.ts).
