+++ okf-bundle/namespace-api-removal-work-queue.md
@@ -8,7 +8,7 @@ timestamp: 2026-06-26T00:00:00Z
 
 # Namespace API removal — work queue
 
-> **IN PROGRESS (2026-06-26):** Latest: `refactor(in-app-messaging): remove deprecated namespaced API`. **Next pickup:** **N3** `messaging` spike.
+> **IN PROGRESS (2026-06-26):** Latest: `refactor(installations): remove deprecated namespaced API`. **Next pickup:** **N3** `messaging` spike.
 > **Order:** pilot smallest (`ml`, `in-app-messaging`) → spike hardest (`messaging`) → bulk small→large → **NF** app cleanup → **NV** full validation. **Workflow:** [namespace-api-removal-workflow.md](namespace-api-removal-workflow.md).
 
 ---
@@ -69,7 +69,7 @@ Update immediately after each work type closes a gate ([fields](testing/iteratio
 | N1 | `ml` (0m) | **closed** | **closed** | **closed** | `refactor(ml): remove deprecated namespace APIs` | — | area-focused | pilot modular-only |
 | N2 | `in-app-messaging` (3m) | **closed** | **closed** | **closed** | `refactor(in-app-messaging): remove deprecated namespaced API` | — | area-focused | web stub macOS; modular e2e 4×3 |
 | N3 | `messaging` (~28m) | open | open | open | — | `gap-analysis` | area-focused | spike |
-| N4 | `installations` (4m) | open | open | open | — | `gap-analysis` | unit-focused | |
+| N4 | `installations` (4m) | **closed** | **closed** | **closed** | `refactor(installations): remove deprecated namespaced API` | — | area-focused | ios/android 4 passing |
 | N4 | `app-distribution` (4m) | open | open | open | — | `gap-analysis` | unit-focused | |
 | N4 | `functions` (4m) | open | open | open | — | `gap-analysis` | unit-focused | |
 | N5 | `perf` (6m) | open | open | open | — | `gap-analysis` | area-focused | |
